Add Sonic Pi (#45)
* Add Sonic Pi

Sonic Pi is a coding synth that has a simple language optimal for teaching programming for beginners, but also a prefectly good tool for professional music producers. It has a really detailed and awesome [description](https://github.com/samaaron/sonic-pi/blob/master/README.md), which I won't even try to beat in this short message about it. I really enjoy playing around with it in my free time.

* Added right format

Failed to get it right the first time